ai-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
spacy==3.7.2
nltk==3.8.1
language-tool-python==2.8
textblob==0.17.1
gensim==4.3.2
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
